+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -418,7 +418,7 @@ class CreateDatasourceResponseBody(BaseModel):
     metadata: CreateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY532EMAA2G1CNYR5A082RY8"
+        "01KY56KEB7B0463PSCE7M35HJ3"
     )
     r"""The unique identifier of the data source"""