+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponseragas.py
@@ -59,9 +59,9 @@ class EvaluatorResponseRagas(BaseModel):
 
     model: str
 
-    created: Optional[str] = "2026-07-22T14:18:06.657Z"
+    created: Optional[str] = "2026-07-22T15:19:38.321Z"
 
-    updated: Optional[str] = "2026-07-22T14:18:06.657Z"
+    updated: Optional[str] = "2026-07-22T15:19:38.321Z"
 
     updated_by_id: OptionalNullable[str] = UNSET