+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -247,7 +247,7 @@ class ListDatasourcesData(BaseModel):
     metadata: ListDatasourcesMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY532EM304ADPDQ39YT12YE0"
+        "01KY56KEB2VQAS4S3SX9WVV735"
     )
     r"""The unique identifier of the data source"""