+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -169,7 +169,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     metadata: RetrieveDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY532EM8PCQDS8P9JZZSWMBT"
+        "01KY56KEB55WZS6PJ4D0X4CYB7"
     )
     r"""The unique identifier of the data source"""