+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -183,7 +183,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     metadata: UpdateDatasourceMetadata
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KY532EMFHJZYT7QRFV7G5VN3"
+        "01KY56KEBCERJAHBGEMSP7DPHZ"
     )
     r"""The unique identifier of the data source"""