+++ packages/orq-rc/RELEASES.md
@@ -9669,3 +9669,13 @@ Based on:
 - [python v4.12.0-rc.55] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.55] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.55 - packages/orq-rc
+
+## 2026-07-22 15:22:28
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.3 (2.918.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.56] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.56] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.56 - packages/orq-rc